docs: document special meaning of registry.npmjs.com
This behavior has been present in npm for a while, but I haven't found
where it's documented. This is my attempt at documenting the behavior
based on my understanding of it. I think a SME should contribute to this
so the documentation is correct.

npm/feedback#544
npm/cli#3783
https://github.com/npm/arborist/blob/478871bf0a44a8ec516b9057585b8707e60b0349/lib/arborist/reify.js#L687-L693

PR-URL: https://github.com/npm/cli/pull/3784
Credit: @everett1992
Close: #3784
Reviewed-by: @wraithgar