Fix --lockfile-version to handle string number
As all CLI input is considered to be string, eg. a "npm install --lockfile-version 3" otherwise fails with the error messages:

npm WARN invalid config lockfile-version="3" set in command line options
npm WARN invalid config Must be one of: null, 1, 2, 3

Replacing `1, 2, 3` with `Number` avoids that and eg. `--lockfile-version 6` will still fail later on with `npm ERR! Invalid lockfileVersion config: 6`, so this is in a way a more DRY approach as well.