chore: update to latest eslint and linting rules
This brings us in line with the rest of the linting rules we are wanting
to use on the npm cli repos.

There were several hundred over-length lines and instead of editing them
all by hand I piped the failing file through `prettier` and back through
`eslint` just to save some time and energy.  This means there may be
some quirks in the linting those files have, but we can fix those if we
see them and they bother us.

Other than that there were about 50 lines that are legitimately
over-length, all are now explicitly overridden.  Many are tests that
could be snapshots.

PR-URL: https://github.com/npm/cli/pull/3995
Credit: @wraithgar
Close: #3995
Reviewed-by: @lukekarrys