fix: dont warn on error cleaning individual log files
Closes: #4128

This also refactors the logic for cleaning log files to use the `ignore`
option from `glob` to not clean current log files instead of relying on
updating the `logs-max` count.

I also discovered a bug where we weren't cleaning log files written with
the old naming convention, so this fixes that as well.

PR-URL: https://github.com/npm/cli/pull/4134
Credit: @lukekarrys
Close: #4134
Reviewed-by: @wraithgar