Change the default save-prefix to ''.
This patch changes the default save-prefix from '^' to ''.

At the moment, there are some massive security implications of using '^'
as the default save directive. Most people will not change this default,
and that means that any dependency of yours can push a malicious PATCH
version, and you will automatically update to it when you next call npm
install.

There are no benefits to having ^ be the default, only security issues.
Contrary to popular belief, it does not mean that you automatically opt
into security patches.