filterSet on packageName
The filterSet is a Set of Node objects, if it is provided in the options for
AuditReport then only nodes contained in that set (and their dependents,
because we use this information to figure out how we might be able to fix any
findings) will be audited.

We'll want to keep any existing filters, so if options.filterSet is already
defined then we'll want to loop through that set and remove any nodes where
!options.package.includes(node.packageName)

If we don't already have a filterSet though, we'll have to create one by
finding nodes with the names found in options.package, which will take
care of anything in the public registry.

Co-authored-by: nlf <quitlahok@gmail.com>