fix: remove lib/utils/read-package-name.js
This code wasn't doing anything special, just dereferencing `name` from
a packument.  There is no need for this to exist.

Most of the tests were able to handle having this go away, except for
`npm owner` which had to have its tests rewritten to be real, which of
course surfaced bugs along the way of behavior that was incorrectly
being tested.  `npm owner` needs some love to clean up its UX, it throws
or returns inconsistently.  I did fix it so that if there is no
package.json in cwd it errored as expected instead of throwing `ENOENT`
which is what it did before.