fix: remove dedupe --save
* Removed dedupe --save documentation and attempted implementation.
* Remove some unneeded otplease mocks from test

`npm dedupe --save` didn't work in a easy to understand way. It would
only update a top level dependency that was duplicated in the tree.
Found this out rewriting the dedupe tests to be real.  This is not very
intuitive and it's best if folks use update or install for saving to
package.json.