chore(changelog): dont show commit body by default & add extra pr check
We don't usually include the commit body in the release notes, so this
removes it from the output by default. It's now behind the
`--format=long` flag, and the command can be rerun with that in case we
need the full commit body sometimes.

Also do an extra check for PR references in titles, since squashed PRs
don't always include the `associatedPullRequests`.