fix: start consolidating color output
chalk already has a way to disable color output, so if we don't want
color we can disable it there and always use that instance of chalk.

This was only updated in the two commands that have real tests.  Doing
it in the other places is going to require making their tests real so
that we don't ALSO have to rewrite their tests just to change their
internal code.