chore(benchmark): add comment permission check and filter paths (#4766)
The updates the issue comment check in the benchmark script, so it only
will trigger if posted by a user with admin rights to the repo. It also
changes the comment to trigger it to `@npm-robot benchmark this`

For pull requests, it now limits benchmarks only when the files changed
include something inside `lib/`.

Requires https://github.com/npm/benchmarks/pull/30 to be merged in the
benchmark repo to allow `issue_comment` as a valid trigger.