fix: disable progress bar on publish
It is not supposed to be there, in that it doesn't get any updates and
gets in the way of logging messages.  We already log the server we are
publishing to in the `notice` headers so the one `http` log message that
we get during publish isn't needed on stdout.