fix: link.target setter
There were guards in place to protect when setting a promise as a
target, which is not something the code actually does, which was
discovered after unpacking load-actual.

This removes those guards which are dangerous because either an
attribute is a promise or it's not.  Letting things access attributes as
non-promises that are sometimes promises is dangerous.