feat: remove `npm bin` (#5459)
BREAKING CHANGE: this removes the `npm bin` command

The output of this command is misleading and incomplete.  The `.bin`
resolution of npm is much more nuanced than this command implies, and
the output of `npm bin` is not something end users should be dealing
with.  `npm` itself is responsible for running the `bin` entries of
modules, with the exception of global bins, which end up in the same
folder as `node` itself, presumably already in a user's path since they
can run node.

Closes https://github.com/npm/statusboard/issues/537