feat: implement new `npm-packlist` behavior
This also lands the latest `pacote` which now requires passing in an
`Arborist` constructor for use in loading the package tree that gets
passed to `npm-packlist`.

BREAKING CHANGE: `npm pack` now follows a strict order of operations
when applying ignore rules. If a files array is present in the
package.json, then rules in .gitignore and .npmignore files from the
root will be ignored.