chore: use packlist to write node_module ignores (#5363)
The gitignore file inside `node_modules/` is now an allow list built
from the output of `npm-packlist`. This means that the only dependencies
that get checked in to source control are those that will end up in the
packed tarball, with the exception of some files that we always ignore.

This also moves the dependencies required to build the docs workspace
from devDependencies to dependencies so they are available for making
the docs after pruning. Ignoring these from source control was the
reason that the change to how we build the gitignore file was necessary.