chore: manually hoist latest versions of some of our deps
I manually installed `@npmcli/fs` and `minipass-fetch` to the root of
our the dependency tree so that the deduped version would now live at
the root of `node_modules/` and any conflicting versions would be deduped
inside of its nested parent `node_modules/` directory. Once this was
locked in `package-lock.json` removing them from the `package.json` does
not undo the hoisting and deduping.

This has no effect on the resolved versions bundled with `npm` but it
does make it easier to visually scan the output of `query` commands to
be sure we are not inadvertently deduping dependencies in the future.