chore: move doc deps to dev
This moves all the dependencies of the `docs/` workspace to dev deps. I
had originally moved them out of devDeps as part of #5309, but this
seems to go against the grain of other tooling. We would have to special
case both `audit` and `licensee` to run on a custom subset of our
dependency tree.

Maybe one day when all tools can be piped the output of an `npm query`
this will be possible. It would be nice if we could only audit
dependencies that are bundled as part of the CLI.

But for now it's easier to move these to devDeps and then reinstall only
the docs workspace after pruning during the publish step.