chore: convert all test to `mockNpm`
This also uses more consistent snapshot cleaning where possible.

Note that this is an iterative step for our tests and does not implement
`MockRegistry` for them, so many of them still mock important
functionality that should be tested. This refactor to `mockNpm` should
make it easier to adopt `MockRegistry` as we see fit.