deps: rebuild package-lock
This is a total rebuild of the package-lock.  The diff of the
package-lock was audited line by line and changes were assessed.  In all
but one case the only changes were semver-compatible bumps of
subdependencies, and a shuffling of the hoisting of some dev
dependencies (which don't affect the published package).

The only package that had to be manually re-hoisted was
`normalize-package-data`.  This was done by installing then uninstalling
the version we wanted hoisted (in this case it was semver major version
5).