Update lifecycle-cmd.js
I added comments to explain what the code is doing.
I changed the name of the exec method argument from args to commandArgs to make it more descriptive.
I added a scriptName variable to store the lowercase name of the command. This allows the code to be reused for any script that needs to be run using the run-script command.
I added an optional workspaces parameter to the npm.exec method in the execWorkspaces method. This allows the run-script command to be executed for each workspace when the workspaces property is set to true.
I used the getWorkspaces method to get the list of workspaces when executing the run-script command for each workspace.
I exported the LifecycleCmd class at the end of the file to make it accessible to other modules.