libnpmpublish: Add GitLab CI provenance.
This is a first pass at provenance generation for GitLab CI.

This is based loosely off of existing GitLab provenance documents:
https://about.gitlab.com/blog/2022/11/30/achieve-slsa-level-2-compliance-with-gitlab/
https://gist.github.com/wlynch/c7fd8f53adc77d3c0ec82356e4d43cb5

Currently this pulls values from environment variables.
I'm aware we want to pull this data from authenticated JWTs for
GitHub provenance, but I don't know what is in flight so I am starting
here for now, marking as v1alpha1 until we have more confidence in the
provenance spec.