Remove redundant and inaccurate statement about the "package.json#files" field
The statement:

'Files included with the "package.json#files" field cannot be excluded through `.npmignore` or `.gitignore`'.

Is inaccurate and redundant as the concept is already being explained right above:

'You can also provide a `.npmignore` file in the root of your package or in subdirectories, which will keep files from being included. At the root of your package it will not override the "files" field, but in subdirectories it will. The `.npmignore` file works just like a `.gitignore.` If there is a `.gitignore` file, and `.npmignore` is missing, `.gitignore`'s contents will be used instead'.