chore: allow publishing new major of workspace without prerelease
Previously the publish script assumed that a new major version of a
workspace would always be preceeded by a prerelease. This updates the
script so that a workspace will be published to the `latest` tag if it
is greater than the current latest major version.

This also fixes an issue where getting a version of a workspace could
fail if it is the first time being published to a new tag. The script
now catches this error and treats it like a workspace that needs
publishing.