feat: save exactly what the user asked for
BREAKING CHANGE: "npm install" will now save the exact range the user gives if one is given.

When adding a new dep to the tree, if a range is given that exact range
is what will be saved to your package.json.  If a dist-tag is used, that
version will be saved with `save-prefix`.  A bare package name is
treated like the dist-tag `latest`.

This also means that `npm update --save` is no longer possible, as that
feature was predicated off of the old behavior of always choosing the
save-prefix and the current version being saved.