chore: correct publish script tag/version logic
The publish script will now never attempt to publish a package that already exists.
The tag to publish to is now only figured out after determining if the package already exists on the registry instead of before.
This also added a confirmation prompt with a full table of name/version/tag of everything that will be published