fix: add back `bin/node-gyp-bin/node-gyp` files
This was an unintended breaking change as part of #6554. The `node-gyp`
bin files are not used by the CLI directly but they are relied on by
other tooling. This change is only intended to land on the v9 release
line.

This partially reverts commit 3a7378d889707d2a4c1f8a6397dda87825e9f5a3.