fix: Remove table output from many commands
This removes table output from `hook`, `profile`, `doctor`, `org`, and `token`.
It also removes table output from the `--long` or `--dry-run` output of
install commands.

Table output is discouraged in a cli for accessibility reasons.

The tests for `token` were also rewritten because they did not actually
test the behavior of npm with the registry.

`this.config` was also removed from `token`.  npm-registry-fetch pulls
all this from flatOptions already.