fix: remove doctor log for each file permission check (#7438)
The previous change was made as part of the removal of the old progress
bar for doctor. Log messages for each individual file scanned were moved
from showing up in the progress message only to `log.silly`. There can
be a very large amount of files scanned, and there is no reason to log
each one individually.