feat: add spinner (#7432)
Closes #7425

This is a little hard to test because everything should continue to work
without progress, as evidenced by the lack of churn in the tests and
snapshots here. The only existing tests that changed are the addition of
newlines after prompts which is new behavior as part of this PR.

I resorted to manually running some commands to get an idea of how the
various states worked together:

**`node . exec -- npm@4`**
This should show progress at the start and then hide it as it prompts
the user to install `npm@4`. `Ctrl+C` should exit from the prompt and
the error should display on the next line (this is a current bug).

**`node . audit signatures --loglevel=http`**
This should show a lot of http log messages while always keeping the
spinner on the last line of output. The spinner also should not jump
between frames regardless of how quickly the log messages show up.

**`node . pack`**
This should immediately show the banners and output from `prepack` and
not show the spinner until the actual packing is happening.

**`node . login`**
The spinner should never while the prompt is being displayed.

**`node . view npm`**
The spinner should appear while data is being fetched and then disappear
for the rest of the command as output is being displayed. The end output
on completion should not have any spinner frames rendered on new lines
in the output. The old progress bar achieved this by calling
`npmlog.disableProgress()` but now it is able to hide the spinner
appropriately even while outputting individual lines to stdout.