fix: limit packument cache size based on heap size
This adds a new packument cache that is an instance of `lru-cache`.
It uses that package's ability to limit content based on size, and has
some multipliers based on research to mostly correctly approximate the
correlation between packument size and its memory usage.  It also limits
the total size of the cache based on the actual heap available.

Closes: https://github.com/npm/cli/issues/7276
Related: https://github.com/npm/pacote/pull/369