chore: disable progress on shellout exit tests
These tests assert what happens if a shellout command like exec throws unexpected errors by
checking what is written to stderr. Progress also gets written to stderr but is not always
deterministic due to the nature of calling it via setInterval. So this disables progress for
these tests so the stderr assertions can be relied on to be the same.