fix: dont write log file for completion commands (#7497)
Since npm completion is implemented by calling npm completion -- arg1
this will write a logfile for that command. I noticed this as I was trying
to look through my logs for a recent command and found that it had been
removed since I used enough tab completions to fill logs-max.