chore: create single bug report issue template (#7505)
This removes the v9 bug report template and changes the v10 template to
specifically mention that it is for bugs for the latest release line.

Previously we would bump these issue templates whenever we did a major
release. I think it makes more sense to have only a single bug template
most of the time and to add templates for old versions when we are
offering bug fix support for them. Currently we don't offer bug fixes
for the v9 release line.

It also removes the `Release 10.x` label which I find to mostly add
noise.
I think it would be better to have no default release line label, and
and instead manually add this label when we identify a bug that only
affects a specific release line that we intend to fix. Or if we are
supporting multiple versions then we can add the label to the bug
template
for old versions.